pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mp-commons"
version = "0.1.0"
description = "Platform shared library – DDD/Hexagonal building blocks for Python services"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
keywords = ["ddd", "hexagonal", "cqrs", "outbox", "observability", "resilience"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Typing :: Typed",
]

# Kernel-only dependencies (intentionally minimal)
dependencies = []

[project.optional-dependencies]
pydantic = ["pydantic>=2.6"]
structlog = ["structlog>=24.1"]
otel = [
  "opentelemetry-api>=1.24",
  "opentelemetry-sdk>=1.24",
  "opentelemetry-exporter-otlp>=1.24",
]
tenacity = ["tenacity>=9.0"]
crypto = ["cryptography>=42.0"]
dotenv = ["python-dotenv>=1.0"]

# Adapter extras (concrete implementations)
fastapi = [
  "fastapi>=0.110",
  "starlette>=0.36",
  "mp-commons[otel]",
]
sqlalchemy = [
  "sqlalchemy>=2.0",
  "alembic>=1.13",
]
redis = ["redis>=5.0"]
kafka = ["aiokafka>=0.10"]
nats = ["nats-py>=2.6"]
rabbitmq = ["aio-pika>=9.4"]
httpx = ["httpx>=0.27"]
requests = ["requests>=2.31"]
keycloak = ["python-jose[cryptography]>=3.3", "httpx>=0.27"]
vault = ["hvac>=2.1"]

# Meta-extras
all-adapters = [
  "mp-commons[fastapi,sqlalchemy,redis,kafka,httpx,keycloak,vault,otel]",
]
dev = [
  "mp-commons[all-adapters,pydantic,structlog,tenacity,crypto,dotenv]",
  "pytest>=8.1",
  "pytest-asyncio>=0.23",
  "pytest-xdist>=3.5",
  "hypothesis>=6.100",
  "coverage[toml]>=7.4",
  "respx>=0.21",
  "time-machine>=2.14",
]

[project.urls]
Homepage = "https://github.com/marcusPrado02/python-commons"
Repository = "https://github.com/marcusPrado02/python-commons"
"Bug Tracker" = "https://github.com/marcusPrado02/python-commons/issues"

# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src/mp_commons"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-ra", "--strict-markers", "--tb=short"]
markers = [
  "unit: fast, in-process tests",
  "integration: tests that hit real infrastructure",
  "contract: schema/contract tests",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 85
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "@(abc\\.)?abstractmethod",
  "\\.\\.\\.",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true
mypy_path = "src"
